Update dns code to use Unmarshal
